Refactor Mock classes and use pytest
