+++ aimmo-game/simulation/avatar/avatar_manager.py
@@ -24,7 +24,8 @@ class AvatarManager(object):
         del self.avatars_by_id[user_id]
 
     def clear_all_avatar_logs(self):
-        [avatar.clear_logs() for avatar in self.avatars]
+        for avatar in self.avatars:
+            avatar.clear_logs()
 
     @property
     def avatars(self):
