+++ aimmo-game/tests/test_service.py
@@ -6,13 +6,9 @@ from simulation.game_state import GameState
 from simulation.interactables.score_location import ScoreLocation
 from simulation.location import Location
 from simulation.world_map import WorldMap
-
 from .test_simulation.dummy_avatar import MoveEastDummy
 from .test_simulation.maps import MockCell, MockPickup
-
-
-class MockGameState(object):
-    avatar_manager = None
+from .test_simulation.mock_game_state import MockGameState
 
 
 class TestService(TestCase):
@@ -22,7 +18,9 @@ class TestService(TestCase):
     @classmethod
     def setUpClass(cls):
         """ Register the api endpoints """
-        cls.game_api = service.GameAPI(worker_manager=None, game_state=MockGameState())
+        cls.game_api = service.GameAPI(
+            worker_manager=None, game_state=MockGameState(None, None)
+        )
 
     def setUp(self):
         """
