+++ aimmo-game/tests/test_simulation/mock_game_state.py
@@ -1,0 +1,11 @@
+from simulation.game_state import GameState
+
+
+class MockGameState(GameState):
+    turn_count = 0
+
+    def serialize(self):
+        return {"foo": "bar"}
+
+    def get_state_for(self, avatar):
+        return self
