+++ aimmo-game/tests/test_simulation/test_log_collector.py
@@ -1,49 +1,30 @@
-from unittest import TestCase
-
 from simulation.avatar.avatar_wrapper import AvatarWrapper
 from simulation.log_collector import LogCollector
-from simulation.workers.worker import Worker
+from .mock_avatar_manager import MockAvatarManager
+from .mock_worker import MockWorker
 from .mock_worker_manager import MockWorkerManager
 
 
-class MockWorker(Worker):
-    def _create_worker(self):
-        pass
+def test_collect_logs():
+    worker_manager = MockWorkerManager()
+    avatar_manager = MockAvatarManager()
 
+    log_collector = LogCollector(worker_manager, avatar_manager)
+    worker = MockWorker(None, None)
+    worker.log = "Worker test log"
 
-class MockAvatarManager(object):
-    avatars_by_id = {}
+    worker_manager.player_id_to_worker[1] = worker
 
-    def add_avatar(self, player_id):
-        avatar = AvatarWrapper(player_id, None, None)
-        self.avatars_by_id[player_id] = avatar
-        return avatar
+    avatar = AvatarWrapper(None, None, None)
 
-    def get_avatar(self, user_id):
-        return self.avatars_by_id[user_id]
+    avatar_manager.avatars_by_id[1] = avatar
 
+    log_collector.collect_logs(1)
 
-class TestLogCollector(TestCase):
-    def test_collect_logs(self):
-        worker_manager = MockWorkerManager()
-        avatar_manager = MockAvatarManager()
+    assert log_collector.player_logs == "Worker test log"
 
-        log_collector = LogCollector(worker_manager, avatar_manager)
-        worker = MockWorker(None, None)
-        worker.log = "Worker test log"
+    avatar.logs.append("Avatar test log")
 
-        worker_manager.player_id_to_worker[1] = worker
+    log_collector.collect_logs(1)
 
-        avatar = AvatarWrapper(None, None, None)
-
-        avatar_manager.avatars_by_id[1] = avatar
-
-        log_collector.collect_logs(1)
-
-        self.assertEquals(log_collector.player_logs, "Worker test log")
-
-        avatar.logs.append("Avatar test log")
-
-        log_collector.collect_logs(1)
-
-        self.assertEquals(log_collector.player_logs, "Worker test logAvatar test log")
+    assert log_collector.player_logs == "Worker test logAvatar test log"
